requirements.txt
pandas
openpyxl
streamlit
plotly
openai
python-dotenv